refactor(auth): share parseBearer; document integration-token last_used_at
First-pass review follow-ups on #114:

- Extract parseBearer into auth/bearer.ts so the integration guard no
  longer depends backwards on api/clients/routes.ts; re-point both
  importers and relocate its unit test to tests/auth/bearer.test.ts.
- Document that an integration token's last_used_at records the last
  successful authentication (bumped even on a later 403 for missing
  scope), and pin that with a guard test.
- Add guard coverage for an empty Bearer credential and route coverage
  for duplicate-scope rejection.
